Make possible the usage for iOS Platforms
